Time: 0 ms (100%), Space: 10.4 MB (75.65%) - LeetHub

+++ 0349-intersection-of-two-arrays/0349-intersection-of-two-arrays.cpp
@@ -3,8 +3,9 @@ public:
     vector<int> intersection(vector<int>& nums1, vector<int>& nums2) {
          sort(nums1.begin(), nums1.end());
         sort(nums2.begin(), nums2.end());
-        int n1 = (int)nums1.size(), n2 = (int)nums2.size();
+        int n1 = nums1.size(), n2 = nums2.size();
         int i1 = 0, i2 = 0;
+        
         vector<int> res;
         while(i1 < n1 && i2 < n2){
             if(nums1[i1] == nums2[i2]) {
@@ -18,8 +19,8 @@ public:
             else{
                 i1++;
             }
-            while(i1 > 0 && i1 < n1 && nums1[i1] == nums1[i1 - 1]) i1++;
-            while(i2 > 0 && i2 < n2 && nums2[i2] == nums2[i2 - 1]) i2++;
+            while(i1>0 && i1 < n1 && nums1[i1] == nums1[i1 - 1]) i1++;
+            while( i2>0 && i2 < n2 && nums2[i2] == nums2[i2 - 1]) i2++;
         }
         return res;
     }
